CTRP unit tests
