+++ tests/unit/ctrp/source/ccle/DepMap-2019q1-celllines.csv_v2.csv
@@ -1,0 +1,3 @@
+DepMap_ID,CCLE_Name,Aliases,COSMIC_ID,Sanger ID,Primary Disease,Subtype Disease,Gender,Source
+ACH-000011,253J_URINARY_TRACT,253J,,,Bladder Cancer,Carcinoma,,KCLB
+ACH-000026,253JBV_URINARY_TRACT,253J-BV,,,Bladder Cancer,Carcinoma,,
