+++ tests/unit/ctrp/source/ccle/cellline_lookup.tsv
@@ -1,0 +1,6 @@
+ACH-000011	ACH-000011
+253J_URINARY_TRACT	ACH-000011
+253J	ACH-000011
+ACH-000026	ACH-000026
+253JBV_URINARY_TRACT	ACH-000026
+253JBV	ACH-000026
