+++ tests/unit/ctrp/source/ccle/cellline_phenotype_lookup.tsv
@@ -1,0 +1,2 @@
+ACH-000011	Carcinoma
+ACH-000026	Carcinoma
