+++ tests/unit/ctrp/source/ccle/cellline_project_lookup.tsv
@@ -1,0 +1,2 @@
+ACH-000011	Bladder_Cancer
+ACH-000026	Bladder_Cancer
